convert TaskContextMenu to typescript
